refactor client module layering
