+++ client/Cargo.toml
@@ -4,6 +4,9 @@ version = "0.1.0"
 edition = "2021"
 license = "Apache-2.0"
 
+[lib]
+path = "lib.rs"
+
 [[bin]]
 name = "client"
 path = "main.rs"
