+++ client/bootstrap/cli.rs
@@ -1,0 +1,8 @@
+use clap::Parser;
+
+#[derive(Parser, Debug)]
+#[command(author, version, about, long_about = None)]
+pub struct Args {
+    #[arg(short, long, default_value = "config/client.yaml")]
+    pub(crate) config: String,
+}
